refactor: move hw funding models
