+++ app/src/main/java/to/bitkit/viewmodels/TransferViewModel.kt
@@ -32,13 +32,13 @@ import to.bitkit.R
 import to.bitkit.data.CacheStore
 import to.bitkit.data.SettingsStore
 import to.bitkit.ext.amountOnClose
+import to.bitkit.models.HwFundingBroadcastResult
+import to.bitkit.models.HwFundingTransaction
 import to.bitkit.models.Toast
 import to.bitkit.models.TransactionSpeed
 import to.bitkit.models.TransferType
 import to.bitkit.models.safe
 import to.bitkit.repositories.BlocktankRepo
-import to.bitkit.repositories.HwFundingBroadcastResult
-import to.bitkit.repositories.HwFundingTransaction
 import to.bitkit.repositories.HwWalletRepo
 import to.bitkit.repositories.LightningRepo
 import to.bitkit.repositories.TransferRepo
